💡refactor: 관심사 분리 및 폴더 구조 변경
