+++ src/components/common/hint/HintButton.tsx
@@ -8,10 +8,7 @@ const HintButton = ({ hint }: HintButtonProps) => {
 
     return (
         <View className="w-[80%] h-12 flex justify-center items-end mb-12">
-            <TouchableOpacity
-                onPress={() => setClickHint(!clickHint)}
-                className="p-2 bg-gray-300 rounded"
-            >
+            <TouchableOpacity onPress={() => setClickHint(!clickHint)} className="p-2 bg-gray-300 rounded">
                 <Text>힌트</Text>
             </TouchableOpacity>
             {clickHint && <HintView setClickHint={setClickHint} hint={hint} />}
