+++ src/components/common/hint/HintView.tsx
@@ -6,10 +6,7 @@ const HintView = ({ setClickHint, hint }: HintViewProps) => {
     return (
         <View className=" absolute w-full min-h-2 pb-5 rounded bg-gray-400 shadow-sm shadow-black">
             <View className="w-full flex justify-center items-end">
-                <TouchableOpacity
-                    onPress={() => setClickHint(false)}
-                    className=" py-1 pr-2"
-                >
+                <TouchableOpacity onPress={() => setClickHint(false)} className=" py-1 pr-2">
                     <Text>x</Text>
                 </TouchableOpacity>
             </View>
