+++ src/components/common/hint/type.ts
@@ -1,0 +1,8 @@
+export type HintButtonProps = {
+    hint?: string;
+};
+
+export type HintViewProps = {
+    setClickHint: (clickHint: boolean) => void;
+    hint?: string;
+};
