+++ src/hooks/type.ts
@@ -2,10 +2,18 @@ import { StackNavigationProp } from '@react-navigation/stack';
 import { RootStackParamList } from '../../type';
 
 export type UseGitHubRedirectProps = {
-  navigation: StackNavigationProp<RootStackParamList>;
+    navigation: StackNavigationProp<RootStackParamList>;
 };
 
 export type UseProblemListProps = {
-  selectedSector: string;
-  selectedDifficulty: number;
+    selectedSector: string;
+    selectedDifficulty: number;
+};
+
+export type sumbitAnswerProps = {
+    formData: {
+        answerText: string;
+        answerOX: boolean | null;
+        selectedOption: string | null;
+    };
 };
